HomeWork-5 done with troubles
