+++ public/index.html
@@ -24,7 +24,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>React Pockemon Game</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
