+++ src/Components/MenuHeader/MenuHeader.js
@@ -5,7 +5,7 @@ import Navbar from "./Navbar/Navbar";
 
 import classes from './MenuHeader.module.css'
 
-const MenuHeader = () => {
+const MenuHeader = ({bgActive}) => {
     const [isActive, setActive] = useState(null);
 
     const reverseHandler = () => {
@@ -16,10 +16,12 @@ const MenuHeader = () => {
         <div className={classes.MenuHeader}>
             <Menu
                 isActive={isActive}
+                reversHandler={reverseHandler}
             />
             <Navbar
                 isActive={isActive}
                 onHandle={reverseHandler}
+                bgActive={bgActive}
             />
         </div>
     );
