+++ src/Components/MenuHeader/Navbar/Navbar.module.css
@@ -1,4 +1,4 @@
-.root {
+#navbar {
     position: fixed;
     z-index: 9999;
     width: 100%;
@@ -7,11 +7,11 @@
     justify-content: center;
 }
 
-.root.bgActive {
+#navbar.bgActive {
     background: #181d23;
 }
 
-.root .navWrapper {
+#navbar .navWrapper {
     width: 100%;
     display: flex;
     justify-content: space-between;
@@ -20,7 +20,7 @@
     padding: 0 2rem;
 }
 
-.root .brand {
+#navbar .brand {
     font-size: 1.6rem;
     color: #fafafa;
     cursor: default;
