+++ src/Components/PockemonCard/PockemonCard.module.css
@@ -29,8 +29,17 @@
     transform: rotateY(180deg);
 }
 
-.pokemonCard.active{
-    transform: rotateY(0deg);
+.pokemonCard.active {
+    animation: flipFront 1000ms 300ms forwards;
+}
+
+@keyframes flipFront {
+    from {
+        transform: rotateY(180deg);
+    }
+    to {
+        transform: rotateY(0deg);
+    }
 }
 
 .cardFront,
