+++ src/routes/About/index.js
@@ -1,0 +1,12 @@
+import classes from './style.module.css'
+
+const AboutPage = () => {
+
+    return (
+        <div className={classes.bg}>
+            <h1>This is page About the game!</h1>
+        </div>
+    );
+};
+
+export default AboutPage;
