+++ src/routes/About/style.module.css
@@ -1,0 +1,5 @@
+.bg {
+    background: url('../../assets/pika.jpg');
+    background-repeat: no-repeat;
+    background-size: cover;
+}
