+++ src/routes/Contact/index.js
@@ -1,0 +1,12 @@
+import classes from './style.module.css'
+
+const ContactPage = () => {
+
+    return (
+        <div className={classes.bg}>
+            <h1>This is CONTACT PAGE!</h1>
+        </div>
+    );
+};
+
+export default ContactPage;
