+++ src/routes/Game/style.module.css
@@ -1,3 +1,10 @@
+.flex {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+}
+
 .HomeButton, .HomeButton:visited {
        width: 100%;
        max-width: 240px;
