+++ src/routes/Home/style.module.css
@@ -1,10 +1,3 @@
-.flex {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-wrap: wrap;
-}
-
 .App {
     text-align: center;
 }
